Revert "NMS-14476: Negate search terms in event advanced search (#5063)"
This reverts commit eb61bd8e4be80e269dee3e6044e50f063adf8cc0.
