--- docs/modules/operation/pages/events/event-advanced-search.adoc
@@ -1,66 +1,0 @@
-
-[[ga-advanced-event-search]]
-= Advanced Event Search
-
-Often users wish to query the event list with more customizations to narrow down the search.
-The {page-component-title} event detail page has a button to filter your searches based on different criteria.
-
-Choose from the following search criteria:
-
-* Event ID
-** Performs an equals filter on the event ID.
-
-* Node Label Contains
-** Performs a substring filter on the node label.
-** To negate this search, prefix the input text with `!`.
-This returns results that do not contain the provided input text.
-
-* Event Text Contains
-** Performs a substring filter on both the log message and the description.
-** To negate this search, prefix the input text with `!`.
-This returns results that do not contain the provided input text.
-
-* TCP/IP Address Like
-** Performs an `IPLIKE` on the provided IP Address to filter the results.
-** To negate this search, prefix the input text with `!`.
-This returns results that do not satisfy the `IPLIKE` function.
-
-* Node Location
-** Select from the dropdown to filter results to only this Node Location.
-** Choose `Any` to return all results.
-
-* System ID
-** Select from the dropdown to filter results to only this System ID.
-** Choose `Any` to return all results.
-
-* Severity
-** Filter the event list based on the event severity.
-** This is a multi-select checkbox.
-Select the severities you wish to filter by.
-By default they are unselected, which returns all events.
-Selecting one or multiple severities restricts your search results to those selected.
-
-* Service
-** Filter the event list based on the alarm service.
-** This is a multi-select checkbox.
-Select the services you wish to filter by.
-By default they are unselected, which returns all events.
-Selecting one or multiple services restricts your search results to those selected.
-
-* Exact Event UEI
-** Performs an equals filter on the event UEI.
-
-* Relative Time
-** Filter the results based on a relative time from now.
-
-* Sort By
-** Sort the results ascending or descending for the selected field name.
-
-* Number of Events per page
-** The number of events per page shown on the list page.
-
-* Events after
-** Select the timestamp to filter for events after this timestamp.
-
-* Events before
-** Select the timestamp to filter for events before this timestamp.
