+++ opennms-webapp/src/main/java/org/opennms/web/alarm/AlarmUtil.java
@@ -29,14 +29,22 @@
 package org.opennms.web.alarm;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Calendar;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.NoSuchElementException;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 import javax.servlet.ServletContext;
 
+import org.apache.commons.lang.ArrayUtils;
 import org.hibernate.criterion.Order;
 import org.hibernate.criterion.Restrictions;
 import org.opennms.core.utils.InetAddressUtils;
+import org.opennms.core.utils.StringUtils;
 import org.opennms.core.utils.WebSecurityUtils;
 import org.opennms.netmgt.model.OnmsAlarm;
 import org.opennms.netmgt.model.OnmsCriteria;
@@ -80,10 +88,6 @@ import org.opennms.web.alarm.filter.SeverityFilter;
 import org.opennms.web.alarm.filter.SeverityOrFilter;
 import org.opennms.web.alarm.filter.SituationFilter;
 import org.opennms.web.filter.Filter;
-import org.opennms.web.utils.filter.CheckboxFilterUtils;
-import org.opennms.web.utils.filter.FilterTokenizeUtils;
-
-import static org.opennms.web.utils.filter.CheckboxFilterUtils.ARRAY_DELIMITER;
 
 /**
  * <p>Abstract AlarmUtil class.</p>
@@ -102,8 +106,20 @@ public abstract class AlarmUtil extends Object {
     /** Constant <code>ANY_RELATIVE_TIMES_OPTION="Any"</code> */
     public static final String ANY_RELATIVE_TIMES_OPTION = "Any";
 
+    /** Constant <code>POSITIVE_CHECKBOX_VALUE="ON"</code> */
+    public static final String POSITIVE_CHECKBOX_VALUE = "on";
+
     /** Constant <code>NEGATION_PREFIX_SYMBOL="!"</code> */
     public static final String NEGATION_PREFIX_SYMBOL = "!";
+
+    /** Constant <code>ARRAY_DELIMITER=","</code> */
+    public static final String ARRAY_DELIMITER = ",";
+
+    /**
+     * Constant <code>MULTI_CHECKBOX_PATTERN="[a-zA-Z]+-\d+=1"</code>
+     * Matches filter strings such as the format "severity-4=1" or "service-1=1" or similar
+     */
+    private static final Pattern MULTI_CHECKBOX_PATTERN = Pattern.compile("[a-zA-Z]+-\\d+=1");
 
     public static OnmsCriteria getOnmsCriteria(final AlarmCriteria alarmCriteria) {
         final OnmsCriteria criteria = new OnmsCriteria(OnmsAlarm.class);
@@ -226,7 +242,7 @@ public abstract class AlarmUtil extends Object {
 
         Filter filter = null;
 
-        String[] tokenizedFilterString = FilterTokenizeUtils.tokenizeFilterString(filterString);
+        String[] tokenizedFilterString = tokenizeFilterString(filterString);
         String type = tokenizedFilterString[0];
         String value = tokenizedFilterString[1];
 
@@ -312,7 +328,7 @@ public abstract class AlarmUtil extends Object {
             filter = new SituationFilter(Boolean.valueOf(value));
         } else if (type.equals(CategoryFilter.TYPE)) {
             String[] nestedFilterString = findFilterString(allFilters, NegativeCategoryFilter.NESTED_TYPE);
-            if (CheckboxFilterUtils.isCheckboxToggled(nestedFilterString)) {
+            if (isCheckboxToggled(nestedFilterString)) {
                 filter = new NegativeCategoryFilter(value);
             } else {
                 filter = new CategoryFilter(value);
@@ -335,10 +351,39 @@ public abstract class AlarmUtil extends Object {
         }
         for (String thisFilter : allFilters) {
             if (thisFilter.startsWith(filterString)) {
-                return FilterTokenizeUtils.tokenizeFilterString(thisFilter);
+                return tokenizeFilterString(thisFilter);
             }
         }
         return null;
+    }
+
+    /**
+     * <p>tokenizeFilterString</p>
+     *
+     * @param filterString a {@link java.lang.String} object.
+     * @return a {@link java.lang.String}[] object representing the type and value tokenized.
+     */
+    private static String[] tokenizeFilterString(String filterString) {
+        String[] tempTokens = filterString.split("=");
+        try {
+            String type = tempTokens[0];
+            String[] values = (String[]) ArrayUtils.remove(tempTokens, 0);
+            String value = org.apache.commons.lang.StringUtils.join(values, "=");
+            return new String[]{type, value};
+        } catch (NoSuchElementException e) {
+            throw new IllegalArgumentException("Could not tokenize filter string: " + filterString);
+        }
+    }
+
+    /**
+     * <p>isCheckboxToggled</p>
+     *
+     * @param tokenizedFilterString a {@link java.lang.String}[] object representing the type and value tokenized.
+     * @return a {@link java.lang.Boolean}[] representing if the option is toggled.
+     */
+    private static boolean isCheckboxToggled(String[] tokenizedFilterString) {
+        return tokenizedFilterString != null &&
+                StringUtils.equalsTrimmed(tokenizedFilterString[1], POSITIVE_CHECKBOX_VALUE);
     }
 
     /**
@@ -425,7 +470,7 @@ public abstract class AlarmUtil extends Object {
     }
 
     public static List<Filter> getFilterList(String[] filterStrings, ServletContext servletContext) {
-        filterStrings = CheckboxFilterUtils.handleCheckboxDuplication(filterStrings);
+        filterStrings = handleCheckboxDuplication(filterStrings);
 
         List<Filter> filterList = new ArrayList<>();
         if (filterStrings != null) {
@@ -437,5 +482,82 @@ public abstract class AlarmUtil extends Object {
             }
         }
         return filterList;
+    }
+
+    /**
+     * <p>handleCheckboxDuplication</p>
+     *
+     * @param filterStrings array of filter strings
+     * @return an array of filter strings which have checkbox deduplication handled. With multi-select there could be
+     * multiple IDs selected so we try to consolidate it into a single filter string
+     */
+    private static String[] handleCheckboxDuplication(String[] filterStrings) {
+        if (filterStrings == null) {
+            return filterStrings;
+        }
+        Map<String, List<String>> selectedCheckboxes = dedupCheckboxSelections(filterStrings);
+
+        if (selectedCheckboxes.isEmpty()) {
+            return filterStrings;
+        }
+        return replaceCheckboxValues(filterStrings, selectedCheckboxes);
+    }
+
+    /**
+     * <p>dedupCheckboxSelections</p>
+     *
+     * @param filterStrings array of filter strings
+     * @return a Map of Filter type to List of IDs associated with the filter.
+     */
+    private static Map<String, List<String>> dedupCheckboxSelections(String[] filterStrings) {
+        Map<String, List<String>> selectedCheckboxes = new HashMap<>();
+
+        for (String filterString : filterStrings) {
+
+            if (MULTI_CHECKBOX_PATTERN.matcher(filterString).matches()) {
+
+                String[] filterStringHyphenSplit = filterString.split("-");
+                String type = filterStringHyphenSplit[0];
+                String remainder = filterStringHyphenSplit[1];
+                String idStr = remainder.split("=")[0];
+
+                if (selectedCheckboxes.containsKey(type)) {
+                    selectedCheckboxes.get(type).add(idStr);
+                } else {
+                    List<String> checkedIds = new ArrayList<>();
+                    checkedIds.add(idStr);
+                    selectedCheckboxes.put(type, checkedIds);
+                }
+            }
+        }
+        return selectedCheckboxes;
+    }
+
+    /**
+     * <p>replaceCheckboxValues</p>
+     *
+     * @param filterStrings array of filter strings
+     * @param selectedCheckboxes map of multiselect filter to IDs
+     * @return the new filter string array with consolidated multi-select checkbox values
+     */
+    private static String[] replaceCheckboxValues(String[] filterStrings, Map<String, List<String>> selectedCheckboxes) {
+        List<String> collectedFilterList = Arrays.stream(filterStrings).filter(thisFilterString -> {
+            for (String selectedKey : selectedCheckboxes.keySet()) {
+                if (thisFilterString.startsWith(selectedKey + "-")) {
+                    return false;
+                }
+            }
+            return true;
+        }).collect(Collectors.toList());
+
+        for (String selectedKey : selectedCheckboxes.keySet()) {
+            List<String> selectedIds = selectedCheckboxes.get(selectedKey);
+            String joinedIds = String.join(ARRAY_DELIMITER, selectedIds);
+            String checkboxFormattedFilter = String.format("%s=%s", selectedKey, joinedIds);
+
+            collectedFilterList.add(checkboxFormattedFilter);
+        }
+
+        return collectedFilterList.toArray(new String[0]);
     }
 }
