+++ opennms-webapp/src/main/java/org/opennms/web/tags/filters/AbstractFilterCallback.java
@@ -29,7 +29,6 @@
 package org.opennms.web.tags.filters;
 
 
-import org.opennms.core.utils.StringUtils;
 import org.opennms.netmgt.model.OnmsFilterFavorite;
 import org.opennms.web.filter.Filter;
 import org.opennms.web.filter.FilterUtil;
@@ -72,8 +71,7 @@ public abstract class AbstractFilterCallback implements FilterCallback {
      */
     @Override
     public List<Filter> parse(String filterString) {
-        String[] filterParameter = (StringUtils.isEmpty(filterString))
-                ? new String[0] : filterString.split("&amp;");
+        String[] filterParameter = filterString.split("&amp;");
         for (int i=0; i< filterParameter.length; i++) {
             if (filterParameter[i].startsWith("filter=")) {
                 filterParameter[i] = filterParameter[i].replaceFirst("filter=", "");
