--- opennms-webapp/src/main/java/org/opennms/web/utils/filter/FilterTokenizeUtils.java
@@ -1,56 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2002-2022 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2022 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.web.utils.filter;
-
-import org.apache.commons.lang.ArrayUtils;
-
-import java.util.NoSuchElementException;
-
-public class FilterTokenizeUtils {
-
-    private static final String QUERY_PARAM_DELIMITER = "=";
-
-    /**
-     * <p>tokenizeFilterString</p>
-     *
-     * @param filterString a {@link java.lang.String} object.
-     * @return a {@link java.lang.String}[] object representing the type and value tokenized.
-     */
-    public static String[] tokenizeFilterString(String filterString) {
-        String[] tempTokens = filterString.split(QUERY_PARAM_DELIMITER);
-        try {
-            String type = tempTokens[0];
-            String[] values = (String[]) ArrayUtils.remove(tempTokens, 0);
-            String value = org.apache.commons.lang.StringUtils.join(values, QUERY_PARAM_DELIMITER);
-            return new String[]{type, value};
-        } catch (NoSuchElementException e) {
-            throw new IllegalArgumentException("Could not tokenize filter string: " + filterString);
-        }
-    }
-}
